base/requirements/dev.txt
-r ./base.txt

#test
flake8>=3.6.0,<3.7.0
pytest-django>=4.4.0,<5.0.0
pytest-cov>=2.12.1,<2.13.0